Move SidebarButton to shared ui folder
No longer being used by just one page
