+++ bikespace_frontend/src/components/parking-map/parking-map-page/ParkingMapPage.tsx
@@ -13,7 +13,7 @@ import {defaultMapCenter, GeocoderSearch} from '@/utils/map-utils';
 
 import {ParkingMapFilter} from './map-filters/ParkingMapFilter';
 import {Sidebar} from './sidebar/Sidebar';
-import {SidebarButton} from '@/components/dashboard/sidebar-button';
+import {SidebarButton} from '@/components/shared-ui/sidebar-button';
 import {
   ParkingFeatureDescription,
   parkingFirstLayerId,
