+++ bikespace_frontend/src/components/shared-ui/sidebar-button/sidebar-button.module.scss
